refactor: group CLI and automation modules into subfolders
Move cli-args, help, and run orchestration under cli/. Move automation,
discovery, create, spec, and scope-spec under automation/. Keep published
entry shims at root paths for npm/JSR/Deno exports.

Closes #150
