+++ knip.json
@@ -1,6 +1,14 @@
 {
   "$schema": "https://unpkg.com/knip@5/schema.json",
-  "entry": ["__tests__/**/*.test.ts", "__tests__/helpers/**/*.ts"],
+  "entry": [
+    "__tests__/**/*.test.ts",
+    "__tests__/helpers/**/*.ts",
+    "src/automation-paths.ts",
+    "src/automation.ts",
+    "src/cli-args.ts",
+    "src/discovery.ts",
+    "src/help.ts"
+  ],
   "project": ["src/**/*.ts", "scripts/**/*.ts", "__tests__/**/*.ts"],
   "ignoreDependencies": ["bun-types"]
 }
